connectiva/protocols/grpc_protocol.py
import grpc
from typing import Dict, Any
from ..interfaces import CommunicationMethod
from ..message import Message

class GrpcProtocol(CommunicationMethod):
    """
    gRPC communication class.
    """

    def __init__(self, **kwargs):
        self.grpc_address = kwargs.get("endpoint").replace("grpc://", "")
        # Initialize your gRPC channel and stubs here
        self.channel = grpc.insecure_channel(self.grpc_address)
        # self.stub = YourGrpcStub(self.channel)

    def connect(self):
        print(f"Connecting to gRPC server at {self.grpc_address}...")

    def send(self, message: Message) -> Dict[str, Any]:
        print(f"Sending message to gRPC server at {self.grpc_address}...")
        try:
            # Convert message to the appropriate gRPC message format
            # Example: grpc_message = YourGrpcMessage(**message.__dict__)
            # response = self.stub.YourMethod(grpc_message)
            print("Message sent successfully!")
            return {}  # Replace with actual response conversion
        except grpc.RpcError as e:
            print(f"Failed to send message: {e}")
            return {"error": str(e)}

    def receive(self) -> Message:
        # Implement receiving logic if needed
        print("Receiving message from gRPC server...")
        return Message(action="receive", data={})

    def disconnect(self):
        print("Disconnecting from gRPC server...")
        self.channel.close()
